# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Project Management System API",
        "description": "A Laravel-based REST API for managing projects, tasks, comments, books, and products.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://laravel1.test",
            "description": "Laravel Herd server"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "operationId": "1bfaa78d1c2c3848ab8165c5dadcad3e",
                "responses": {
                    "200": {
                        "description": "List of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "description": "Creates a new product with name, price, quantity, and optional description. All fields are validated before creation.",
                "operationId": "9934bcf87d6613c14890dbdd587d311a",
                "requestBody": {
                    "description": "Product data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductRequest"
                            },
                            "example": {
                                "name": "Wireless Mouse",
                                "price": 29.99,
                                "quantity": 50,
                                "description": "Ergonomic wireless mouse with 2.4GHz connectivity"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                },
                                "example": {
                                    "id": 1,
                                    "name": "Wireless Mouse",
                                    "price": "29.99",
                                    "quantity": 50,
                                    "description": "Ergonomic wireless mouse with 2.4GHz connectivity",
                                    "created_at": "2025-11-30T18:00:00.000000Z",
                                    "updated_at": "2025-11-30T18:00:00.000000Z"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name field is required."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/products/{product}/reduce-stock": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Reduce product stock",
                "operationId": "f52c8bf89614d0186bf76f281b14fd1d",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Stock reduction data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReduceStockRequest"
                            },
                            "example": {
                                "amount": 5
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stock reduced successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                },
                                "example": {
                                    "id": 1,
                                    "name": "Wireless Mouse",
                                    "price": "29.99",
                                    "quantity": 45,
                                    "description": "Ergonomic wireless mouse",
                                    "created_at": "2025-11-30T18:00:00.000000Z",
                                    "updated_at": "2025-12-01T11:00:00.000000Z"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Insufficient stock or validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Insufficient stock. Available: 2"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No query results for model [App\\Models\\Product] 999"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "description": "Product model with inventory information",
                "properties": {
                    "id": {
                        "description": "Unique product identifier",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Product name",
                        "type": "string",
                        "example": "Laptop Stand"
                    },
                    "price": {
                        "description": "Product price in decimal format",
                        "type": "string",
                        "format": "decimal",
                        "example": "49.99"
                    },
                    "quantity": {
                        "description": "Available stock quantity",
                        "type": "integer",
                        "example": 100
                    },
                    "description": {
                        "description": "Product description",
                        "type": "string",
                        "example": "Adjustable aluminum laptop stand with ergonomic design",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-30T18:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-11-30T18:00:00.000000Z"
                    }
                },
                "type": "object"
            },
            "StoreProductRequest": {
                "required": [
                    "name",
                    "price",
                    "quantity"
                ],
                "properties": {
                    "name": {
                        "description": "Product name",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Wireless Mouse"
                    },
                    "price": {
                        "description": "Product price (must be >= 0)",
                        "type": "number",
                        "format": "float",
                        "minimum": 0,
                        "example": 29.99
                    },
                    "quantity": {
                        "description": "Initial stock quantity (must be >= 0)",
                        "type": "integer",
                        "minimum": 0,
                        "example": 50
                    },
                    "description": {
                        "description": "Optional product description",
                        "type": "string",
                        "example": "Ergonomic wireless mouse with 2.4GHz connectivity",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ReduceStockRequest": {
                "required": [
                    "amount"
                ],
                "properties": {
                    "amount": {
                        "description": "Amount to reduce from stock (must be >= 1)",
                        "type": "integer",
                        "minimum": 1,
                        "example": 5
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}